shark/app/server/requirements.txt
MySQL-python==1.2.5
Django==1.9.0
djangorestframework==3.4.0
